split basic.py into subfiles in module
