+++ numth/basic/__init__.py
@@ -1,0 +1,37 @@
+#   numth/basic/__init__.py
+#===========================================================
+
+from .division import (
+    bezout,
+    div,
+    div_with_small_remainder,
+    euclidean_algorithm,
+    gcd,
+    lcm,
+    padic
+)
+
+from .modular import (
+    chinese_remainder_theorem,
+    euler_criterion,
+    jacobi,
+    mod_inverse,
+    mod_power,
+    prime_to
+)
+
+from .primality import (
+    is_prime__naive,
+    prime_sieve
+)
+
+from .shape_number import (
+    shape_number_by_index,
+    which_shape_number
+)
+
+from .sqrt import (
+    integer_sqrt,
+    is_square
+)
+
